refactor: extract Clock to a stand alone package
