--- worker/clock.go
@@ -1,15 +1,0 @@
-package worker
-
-import "time"
-
-type Clock interface {
-	Now() time.Time
-}
-
-type defaultClock struct{}
-
-func (defaultClock) Now() time.Time {
-	return time.Now()
-}
-
-var DefaultClock = defaultClock{}
