+++ worker/mocks.go
@@ -1,8 +1,6 @@
 package worker
 
 import (
-	"time"
-
 	pb "github.com/raystack/raccoon/proto"
 	mock "github.com/stretchr/testify/mock"
 )
@@ -28,12 +26,4 @@ type mockAck struct {
 
 func (m *mockAck) Ack(err error) {
 	m.Called(err)
-}
-
-type mockClock struct {
-	mock.Mock
-}
-
-func (m *mockClock) Now() time.Time {
-	return m.Called().Get(0).(time.Time)
 }
